+++ src/Predicate.php
@@ -43,6 +43,24 @@ class Predicate
     */
     public function add($logic,$arg)
     {
+        if (!is_array($arg))
+        {
+            throw new \InvalidArgumentException('Predicate Error: argument passed must be type of array');
+        }
+
+        if (count($arg) == 1)
+        {
+            if (isset($arg[0]) && is_array($arg[0]))
+            {
+                foreach($arg[0] as $key => $value)
+                {
+                    if ($value == '') continue;
+
+                    $arg = $this->formatWhere($key, $value);
+                }
+            }
+        }
+
         if (count($arg) != 3)
         {
             throw new \InvalidArgumentException('Predicate Error: Must have 3 arguments passed - '.count($arg).' given');
@@ -61,6 +79,18 @@ class Predicate
         }
 
         $this->predicates[$logic][] = $arg;
+    }
+
+
+    //--------------------------------------------------------------------
+
+    /**
+    * formatWhere
+    *
+    */
+    protected function formatWhere($key, $value)
+    {
+        return [$key,'==',$value];
     }