+++ src/Query.php
@@ -109,36 +109,7 @@ class Query extends QueryLogic
     */
     protected function addPredicate($logic,$arg)
     {
-        if (count($arg) == 3)
-        {
-            $this->predicate->add($logic, $arg);
-        }
-
-        if (count($arg) == 1)
-        {
-            if (isset($arg[0]) && count($arg[0]))
-            {
-                foreach($arg[0] as $key => $value)
-                {
-                    if ($value == '') continue;
-
-                    $this->predicate->add($logic, $this->formatWhere($key, $value));
-                }
-            }
-        }
-    }
-
-
-    //--------------------------------------------------------------------
-
-
-    /**
-    * formatWhere
-    *
-    */
-    protected function formatWhere($key, $value)
-    {
-        return [$key,'==',$value];
+        $this->predicate->add($logic, $arg);
     }
 
 
@@ -182,6 +153,21 @@ class Query extends QueryLogic
 
 
     //--------------------------------------------------------------------
+
+
+    /**
+    * ->first()
+    *
+    */
+    public function first()
+    {
+        $results = parent::run()->toArray();
+        return current($results);
+    }
+
+
+    //--------------------------------------------------------------------
+
 
 
     /**