+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+### 08/05/2017 - 1.0.4
+* Added `first()` (if you want to only return the first array of the query result)
+* Ability to use Queries without needing `where()`, can now use queries to find all and order results
+* Fixed Predicate Exceptions for bad query arguments (now correctly parsing them)
+
 ### 08/05/2017 - 1.0.3
 * Added `orderBy()` (sorting field and direction `ASC` and `DESC`)
 * Added `limit()` Limit results returned, includes Limit and Offset options.